+++ app/src/main/java/nextstep/signup/model/validation/CompositeValidation.kt
@@ -1,4 +1,4 @@
-package nextstep.signup.ui.validation
+package nextstep.signup.model.validation
 
 class CompositeValidation(
     private vararg val validations: Validation
