+++ app/src/main/java/nextstep/signup/model/validation/Validation.kt
@@ -1,4 +1,4 @@
-package nextstep.signup.ui.validation
+package nextstep.signup.model.validation
 
 interface Validation {
     fun validate(text: String): ValidationResult
