+++ app/src/test/java/nextstep/signup/EmailValidationTest.kt
@@ -4,7 +4,7 @@ import nextstep.signup.ui.signup.EmailTextField
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
-import nextstep.signup.ui.validation.RegexValidation
+import nextstep.signup.model.validation.RegexValidation
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
