+++ app/src/test/java/nextstep/signup/PasswordConfirmValidationTest.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import nextstep.signup.ui.signup.PasswordConfirmTextField
-import nextstep.signup.ui.validation.EqualValidation
+import nextstep.signup.model.validation.EqualValidation
 import nextstep.signup.ui.signup.PasswordTextField
 import org.junit.Before
 import org.junit.Rule
